feat: Support a TensorFlow style RMSProp optimizer.
The RMSProp impelentation in TensorFlow is different from the Pytorch one:
 - Momentum is applyed to learning rates as well as gradients.
 - Machine epsilon is added before applying square root.
Beacuse of above, the hyper parameters for RMSProp in Pytorch must be
different from the parameters in TensorFlow.
To solve this issue, a TensorFlow style RMSProp is implemented as `rmsproptf`.

The new RMSProps is imported from "EfficientNetV2-S implementation using PyTorch":
https://github.com/jahongir7174/EffcientNetV2.