fix: Fix the architecture of Skip-EfficientNets and Skip-MobileNets.
EffifientNets and MobileNets do not have downsample modules,
but Skip-EfficientNets and Skip-MobileNets needs the donwsample modules.
In the current version, Skip-EfficientNets and Skip-MobileNets have
the downsample modules, but they are removed from EfficientNets and MobileNets.
In addition, downsample modules are removed from Skip-SwinTransformers,
because SwinTransformers do not require any downsample modules.