refactor: Change Block class to support pre-downsample architecture.
In the previous version, Block class have supported only ResNets
style downsamples, but the Block class is updated to support
pre-downsample architecture such as Swin or ConvNeXt in the
current version. Because of this change, argument of Block class is
changed, and weight parameters of Swin and ConvNeXt are also changed.